Fix code tags in desktop metadata
While the spec officially supports `<em>` and `<code>`-tags, the
libraries used by the flathub frontend as well as GNOME software, don't
support these tags and just strip them, resulting in a non-functional
description for the flatpak on multiple platforms. While the right way
to do it, would be to fix to be applied to the upstream version of the
appstream-glib library, this doesn't seem to happen any time soon.

Therefore this patch removes the usage of the `<code>` and replaces it
with supported `<p>` tags, which might result in broken syntax if the
fonts substitute characters. That's nothing we can change.

References:
https://github.com/guinux/appstream-glib/commit/7fef9b4e0f2bbb33bd0f036413f3523ca5e821b8
https://github.com/flathub/linux-store-frontend/issues/203