fix: Add workaround to disable broken search
This patch intentionally breaks search and can be used to disable the
Element Search indexing for users unable to navigate to the right
settings page on time.

This is achieved by replacing the existing search index directory
(`${XDG_CONFIG_HOME}/<Element Profile>/EventStore`) with a file,
triggering EEXIST error and resulting in a failed initialisation. This
can easily be fixed using the UI and as long as all messages are still
available on the homeserver, no data is lost and the index can be
rebuilt.

In order to use this workaround, either start the flatpak using `flatpak
run --env=DISABLE_ENCRYPTED_SEARCH=true im.riot.Riot` or add the
environment variable using flatseal. A one time treatment should be
enough as long as you already have a profile with an EventStore
directory.